telegram-gpt5-pdf-qa-bot/requirements.txt
python-telegram-bot==22.3
openai>=1.41.0
pypdf>=4.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
